fix: 테스트데이터 셋업 메소드 분리
