+++ test/unit/pet/presentation/pet.controller.test.ts
@@ -1,17 +1,11 @@
-import { Builder } from 'builder-pattern';
 import { AuthProvider } from '../../../../src/auth/presentation/user.dto';
 import { PetService } from '../../../../src/pet/application/pet.service';
-import { Pet } from '../../../../src/pet/pet.domain';
 import { PetController } from '../../../../src/pet/presentation/pet.controller';
 import { PetDto } from '../../../../src/pet/presentation/pet.dto';
-import { BreedEntity } from '../../../../src/schemas/breed.entity';
 import { CustomerEntity } from '../../../../src/schemas/customer.entity';
-import { PetChecklistAnswerEntity } from '../../../../src/schemas/pet-checklist-answer.entity';
-import { PetChecklistChoiceEntity } from '../../../../src/schemas/pet-checklist-chocie.entity';
 import {
   ChecklistType,
   PetChecklistCategory,
-  PetChecklistEntity,
 } from '../../../../src/schemas/pet-checklist.entity';
 import { Gender } from '../../../../src/schemas/pets.entity';
 import { FakeCustomerRepository } from '../../../mock/fake.customer.repository';
@@ -22,6 +16,9 @@ import { FakePetChecklistChoiceAnswerRepository } from '../../../mock/pet/fake.p
 import { FakePetChecklistChoiceRepository } from '../../../mock/pet/fake.pet.checklist-choice.repository';
 import { FakePetChecklistRepository } from '../../../mock/pet/fake.pet.checklist.repository';
 import { FakePetRepository } from '../../../mock/pet/fake.pet.repository';
+import createPetDto, {
+  setupInitialData,
+} from '../../../mock/pet/pet.setup-data';
 
 describe('PetController', () => {
   let petController: PetController;
@@ -209,98 +206,3 @@ describe('PetController', () => {
     });
   });
 });
-
-function createPetDto(overrides?: Partial<PetDto>): PetDto {
-  const basePetDto = Builder<PetDto>()
-    .petName('몽글이')
-    .petGender(Gender.MALE)
-    .petBirthdate(new Date())
-    .petWeight(10)
-    .neuteredYn(true)
-    .personality('cute')
-    .vaccinationStatus('미완료')
-    .appointments([])
-    .breedId(1)
-    .build();
-
-  return { ...basePetDto, ...overrides };
-}
-
-export async function setupInitialData(
-  fakeCustomerRepository: FakeCustomerRepository,
-  fakeBreedRepository: FakeBreedRepository,
-  fakePetRepository: FakePetRepository,
-  fakePetChecklistRepository: FakePetChecklistRepository,
-  fakePetChecklistChoiceRepository: FakePetChecklistChoiceRepository,
-  fakePetChecklistAnswerRepository: FakePetChecklistAnswerRepository,
-  fakePetChecklistChoiceAnswerRepository: FakePetChecklistChoiceAnswerRepository,
-  fakeUuidHolder: FakeUuidHolder,
-  fakeDateHolder: FakeDateHolder,
-) {
-  // Customer 초기 데이터 (생성 후 반환값 재사용)
-  const customer: CustomerEntity = await fakeCustomerRepository.create({
-    customerName: '홍길동',
-    authProvider: AuthProvider.APPLE,
-  });
-
-  // Breed 초기 데이터
-  fakeBreedRepository.create(
-    {
-      breedId: 1,
-      breedName: '시고르잡종',
-      breedDescription: '귀엽다',
-      uuid: fakeUuidHolder.generatedUuid(),
-    },
-    fakeUuidHolder,
-  );
-
-  // Pet 초기 데이터
-  const breed: BreedEntity = await fakeBreedRepository.getBreed(1);
-  if (!breed) {
-    throw new Error('Breed not found');
-  }
-
-  const petDto = createPetDto();
-
-  const pet = await fakePetRepository.create(
-    Pet.create(petDto, breed, customer, fakeUuidHolder, fakeDateHolder),
-    customer,
-    fakeDateHolder,
-    fakeUuidHolder,
-  );
-
-  // PetChecklist 초기 데이터
-  const checklist: PetChecklistEntity = await fakePetChecklistRepository.create(
-    {
-      petChecklistId: 1,
-      petChecklistType: ChecklistType.ANSWER,
-      petChecklistCategory: PetChecklistCategory.HEALTH,
-      petChecklistContent: '반려동물의 현재 건강 상태가 어떤가요?',
-    },
-  );
-
-  // PetChecklistChoice 초기 데이터
-  const petChecklistChoice: PetChecklistChoiceEntity =
-    await fakePetChecklistChoiceRepository.create({
-      petChecklistChoiceId: 1,
-      petChecklistChoiceContent: '좋다',
-    });
-
-  // PetChecklistAnswer 초기 데이터
-  const petChecklistAnswer: PetChecklistAnswerEntity =
-    await fakePetChecklistAnswerRepository.create({
-      petChecklistId: checklist.petChecklistId,
-      petId: pet.petId,
-      petChecklistAnswer: 'Yes',
-    });
-
-  // Checklist에 Answer를 연결
-  checklist.petChecklistAnswers = [petChecklistAnswer];
-  await fakePetChecklistRepository.update(checklist.petChecklistId, checklist);
-
-  await fakePetChecklistChoiceAnswerRepository.create({
-    petId: pet.petId,
-    petChecklistId: checklist.petChecklistId,
-    petChecklistChoiceId: petChecklistChoice.petChecklistChoiceId,
-  });
-}
